# api/openapi.yaml
openapi: 3.1.0
info:
  title: Conductor API
  version: 0.6.0
  description: >-
    The server explicitly selects local or oidc authentication. Local mode requires
    X-Conductor-Actor and exposes only unscoped legacy packages. OIDC mode requires
    an RFC 9068 RS256 bearer access token or configured browser session and rejects local actor headers. Signed
    identity is mapped to server-managed PostgreSQL permissions; token role claims
    grant no rights. Workspace selection is required for package and repository
    operations. Inaccessible package IDs return the same 404 as unknown IDs.
servers: [{url: http://localhost:8080}]
security:
  - BearerAuth: []
  - BrowserCookie: []
  - LocalActor: []
paths:
  /api/v1/auth/config:
    get:
      summary: Discover the explicit server authentication mode
      description: Public bounded metadata; no query parameters. An unavailable response must never cause a local-actor fallback.
      security: []
      responses:
        '200': {description: Configured authentication, content: {application/json: {schema: {type: object, required: [mode, browserLogin], properties: {mode: {type: string, enum: [local, oidc]}, browserLogin: {type: boolean}}}}}}
        '400': {$ref: '#/components/responses/BadRequest'}
...
  /api/v1/changes:
    get:
      summary: Discover changes shared through the central store
      description: >-
        Lists each change's latest revision and effective current approval, ordered
        by change creation time descending, then ID descending. Other developers'
        committed packages and revisions are visible through this same store.
        This endpoint does not report active agents or execution. Repository filters
        match the latest content.repositoryContext.repository string exactly;
        malformed or missing repository labels are omitted. Pagination uses immutable
        change creation coordinates; new entries appear when starting a new listing.
        Workspace/repository authorization filters rows before pagination in OIDC
        mode; author-supplied repository labels never establish permissions.
        Unknown, repeated, malformed, or out-of-range query parameters are rejected.
      parameters:
        - $ref: '#/components/parameters/Actor'
        - $ref: '#/components/parameters/Workspace'
        - $ref: '#/components/parameters/Repository'
        - $ref: '#/components/parameters/CSRF'
        - name: repository
          in: query
          description: Exact repository identity; omitted or empty lists all changes; at most 2048 UTF-8 bytes
          schema: {type: string, maxLength: 2048}
        - name: before
          in: query
          description: Opaque continuation cursor from nextBefore; omitted or empty starts a new listing
          schema: {type: string, maxLength: 1024}
        - $ref: '#/components/parameters/HistoryLimit'
      responses:
        '200':
          description: Bounded shared change listing, including an empty list when nothing matches
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ChangePage'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '500': {$ref: '#/components/responses/InternalError'}
        '503': {$ref: '#/components/responses/Unavailable'}
    post:
      summary: Create an initial immutable work-package revision
      description: Accepts no query parameters; identity and scope use the documented headers, and command inputs use the JSON body.
      parameters: [{$ref: '#/components/parameters/Actor'}, {$ref: '#/components/parameters/Workspace'}, {$ref: '#/components/parameters/Repository'}, {$ref: '#/components/parameters/BrowserOrigin'}, {$ref: '#/components/parameters/CSRF'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/ContentCommand'}}}}
      responses: {'201': {description: Created, content: {application/json: {schema: {$ref: '#/components/schemas/Package'}}}}, '400': {$ref: '#/components/responses/BadRequest'}, '401': {$ref: '#/components/responses/Unauthorized'}, '403': {$ref: '#/components/responses/Forbidden'}, '404': {$ref: '#/components/responses/NotFound'}}
  /api/v1/changes/{id}:
    parameters: [{$ref: '#/components/parameters/ID'}, {$ref: '#/components/parameters/Actor'}, {$ref: '#/components/parameters/Workspace'}, {$ref: '#/components/parameters/Repository'}, {$ref: '#/components/parameters/BrowserOrigin'}, {$ref: '#/components/parameters/CSRF'}]
    get:
      summary: Inspect the latest revision and its effective approval
      description: Accepts no query parameters; identity and scope use the documented headers, and command inputs use the JSON body.
      responses: {'400': {$ref: '#/components/responses/BadRequest'}, '200': {description: Found, content: {application/json: {schema: {$ref: '#/components/schemas/Package'}}}}, '401': {$ref: '#/components/responses/Unauthorized'}, '403': {$ref: '#/components/responses/Forbidden'}, '404': {$ref: '#/components/responses/NotFound'}}
  /api/v1/changes/{id}/history:
    parameters: [{$ref: '#/components/parameters/ID'}, {$ref: '#/components/parameters/Actor'}, {$ref: '#/components/parameters/Workspace'}, {$ref: '#/components/parameters/Repository'}, {$ref: '#/components/parameters/BrowserOrigin'}, {$ref: '#/components/parameters/CSRF'}]
    get:
      summary: Inspect revision summaries and retained approval counts
      description: >-
        Returns revisions in descending number order. Approval counts describe
        historical records and do not grant effective approval to the current
        package. Unknown, repeated, empty, or malformed query parameters are rejected.
      parameters:
        - $ref: '#/components/parameters/BeforeRevision'
        - $ref: '#/components/parameters/HistoryLimit'
      responses:
        '200':
          description: Bounded revision history, including an empty page after exhaustion
          content:
            application/json:
              schema: {$ref: '#/components/schemas/HistoryPage'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '500': {$ref: '#/components/responses/InternalError'}
        '503': {$ref: '#/components/responses/Unavailable'}
  /api/v1/changes/{id}/revisions/{revision}:
    parameters:
      - $ref: '#/components/parameters/ID'
      - $ref: '#/components/parameters/Actor'
      - $ref: '#/components/parameters/Workspace'
      - $ref: '#/components/parameters/Repository'
      - $ref: '#/components/parameters/CSRF'
      - name: revision
        in: path
        required: true
        schema: {type: integer, format: int64, minimum: 1}
    get:
      summary: Inspect exact revision content and its retained approval records
      description: >-
        This read-only record has no effective approved flag. Approvals are ordered
        by creation time, then reviewer, and capped at 100. approvalsTruncated is
        true when further approval records exist. Unknown query parameters are
        rejected; this operation accepts none.
      responses:
        '200':
          description: Exact revision and bounded historical approvals
          content:
            application/json:
              schema: {$ref: '#/components/schemas/RevisionRecord'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '500': {$ref: '#/components/responses/InternalError'}
        '503': {$ref: '#/components/responses/Unavailable'}
  /api/v1/changes/{id}/events:
    parameters: [{$ref: '#/components/parameters/ID'}, {$ref: '#/components/parameters/Actor'}, {$ref: '#/components/parameters/Workspace'}, {$ref: '#/components/parameters/Repository'}, {$ref: '#/components/parameters/BrowserOrigin'}, {$ref: '#/components/parameters/CSRF'}]
    get:
      summary: Inspect the change's audit events in sequence order
      description: >-
        Returns events in ascending sequence order. Sequence numbers may contain
        gaps. Unknown, repeated, empty, or malformed query parameters are rejected.
      parameters:
        - $ref: '#/components/parameters/AfterSequence'
        - $ref: '#/components/parameters/HistoryLimit'
      responses:
        '200':
          description: Bounded audit history, including an empty page after exhaustion
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AuditPage'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '500': {$ref: '#/components/responses/InternalError'}
        '503': {$ref: '#/components/responses/Unavailable'}
  /api/v1/changes/{id}/revisions:
    parameters: [{$ref: '#/components/parameters/ID'}, {$ref: '#/components/parameters/Actor'}, {$ref: '#/components/parameters/Workspace'}, {$ref: '#/components/parameters/Repository'}, {$ref: '#/components/parameters/BrowserOrigin'}, {$ref: '#/components/parameters/CSRF'}]
    post:
      summary: Append a revision using optimistic concurrency
      description: Accepts no query parameters; identity and scope use the documented headers, and command inputs use the JSON body.
      requestBody: {required: true, content: {application/json: {schema: {type: object, additionalProperties: false, required: [expectedRevision, content], properties: {expectedRevision: {type: integer}, content: {type: object, additionalProperties: true}}}}}}
      responses: {'201': {description: Revised, content: {application/json: {schema: {$ref: '#/components/schemas/Package'}}}}, '400': {$ref: '#/components/responses/BadRequest'}, '401': {$ref: '#/components/responses/Unauthorized'}, '403': {$ref: '#/components/responses/Forbidden'}, '404': {$ref: '#/components/responses/NotFound'}, '409': {$ref: '#/components/responses/Conflict'}}
  /api/v1/changes/{id}/review-requests:
    parameters: [{$ref: '#/components/parameters/ID'}, {$ref: '#/components/parameters/Actor'}, {$ref: '#/components/parameters/Workspace'}, {$ref: '#/components/parameters/Repository'}, {$ref: '#/components/parameters/BrowserOrigin'}, {$ref: '#/components/parameters/CSRF'}]
    post:
      summary: Submit the exact current revision
      description: Accepts no query parameters; identity and scope use the documented headers, and command inputs use the JSON body.
      requestBody: {required: true, content: {application/json: {schema: {type: object, additionalProperties: false, required: [revision], properties: {revision: {type: integer}}}}}}
      responses: {'200': {description: Submitted, content: {application/json: {schema: {$ref: '#/components/schemas/Package'}}}}, '400': {$ref: '#/components/responses/BadRequest'}, '401': {$ref: '#/components/responses/Unauthorized'}, '403': {$ref: '#/components/responses/Forbidden'}, '404': {$ref: '#/components/responses/NotFound'}, '409': {$ref: '#/components/responses/Conflict'}}
  /api/v1/changes/{id}/approvals:
    parameters: [{$ref: '#/components/parameters/ID'}, {$ref: '#/components/parameters/Actor'}, {$ref: '#/components/parameters/Workspace'}, {$ref: '#/components/parameters/Repository'}, {$ref: '#/components/parameters/BrowserOrigin'}, {$ref: '#/components/parameters/CSRF'}]
    post:
      summary: Approve the exact inspected revision and digest
      description: Accepts no query parameters; identity and scope use the documented headers, and command inputs use the JSON body.
      requestBody: {required: true, content: {application/json: {schema: {type: object, additionalProperties: false, required: [revision, digest], properties: {revision: {type: integer}, digest: {type: string, pattern: '^[a-f0-9]{64}$'}}}}}}
      responses: {'201': {description: Approved, content: {application/json: {schema: {$ref: '#/components/schemas/Package'}}}}, '400': {$ref: '#/components/responses/BadRequest'}, '401': {$ref: '#/components/responses/Unauthorized'}, '403': {$ref: '#/components/responses/Forbidden'}, '404': {$ref: '#/components/responses/NotFound'}, '409': {$ref: '#/components/responses/Conflict'}, '422': {description: Approval rejected}}
  /api/v1/context-collections:
    post:
      summary: Persist one explicit remote context collection request
      description: >-
        Requires OIDC identity, selected workspace and repository, current author
        permission, and an operator-enabled integration. Local or unconfigured
        services return 503. Unknown or duplicate command fields and query parameters are rejected. The exact lowercase
        SHA-1 commit and 1-32 unique paths are immutable; validated paths are sorted
        bytewise before input hashing. A single Idempotency-Key is scoped to the
        requester and canonical repository. Reusing it with the same input returns
        the original request; changed input conflicts. HTTP 202 establishes only
        persisted intent, not a completed collection, passing evidence, or approval.
      security: [{BearerAuth: []}, {BrowserCookie: []}]
      parameters:
        - $ref: '#/components/parameters/CollectionWorkspace'
        - $ref: '#/components/parameters/CollectionRepository'
        - $ref: '#/components/parameters/BrowserOrigin'
        - $ref: '#/components/parameters/CSRF'
        - $ref: '#/components/parameters/CollectionKey'
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/CollectionInput'}}}}
      responses:
        '202': {description: Persisted request or its idempotent replay, content: {application/json: {schema: {$ref: '#/components/schemas/Collection'}}}}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {description: Idempotency key already identifies different input (idempotency_conflict), content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}
        '429': {$ref: '#/components/responses/CollectionCapacity'}
        '500': {$ref: '#/components/responses/InternalError'}
        '503': {$ref: '#/components/responses/Unavailable'}
    get:
      summary: List shared context collection facts under current permissions
      description: >-
        Requires a selected workspace and repository with current read permission.
        Returns bounded summaries without source text, ordered by immutable creation
        time descending then ID descending. Permission filtering occurs before the
        page limit. Unknown, repeated, malformed, or out-of-range query parameters
        are rejected. An execution observation is separate from receipt coverage;
        missing or stale observations never establish current progress.
      security: [{BearerAuth: []}, {BrowserCookie: []}]
      parameters:
        - $ref: '#/components/parameters/CollectionWorkspace'
        - $ref: '#/components/parameters/CollectionRepository'
        - $ref: '#/components/parameters/CSRF'
        - {name: before, in: query, description: Opaque nextBefore cursor; omitted or empty starts a new listing, schema: {type: string, maxLength: 1024}}
        - $ref: '#/components/parameters/HistoryLimit'
      responses:
        '200': {description: Shared collection summaries and explicit continuation, content: {application/json: {schema: {$ref: '#/components/schemas/CollectionPage'}}}}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '500': {$ref: '#/components/responses/InternalError'}
        '503': {$ref: '#/components/responses/Unavailable'}
  /api/v1/context-collections/{id}:
    get:
      summary: Inspect a stored collection, receipt, and execution observation
      description: >-
        Requires current read permission in the selected canonical scope. Unknown
        and inaccessible IDs return the same 404. Historical source remains
        protected after revocation. Accepts no query parameters. Inspection does
        not fetch provider content or refresh a receipt; absence of a receipt
        means no persisted result, not passing evidence.
      security: [{BearerAuth: []}, {BrowserCookie: []}]
      parameters:
        - $ref: '#/components/parameters/CollectionID'
        - $ref: '#/components/parameters/CollectionWorkspace'
        - $ref: '#/components/parameters/CollectionRepository'
        - $ref: '#/components/parameters/CSRF'
      responses:
        '200': {description: Persisted request and available historical observations, content: {application/json: {schema: {$ref: '#/components/schemas/Collection'}}}}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '500': {$ref: '#/components/responses/InternalError'}
        '503': {$ref: '#/components/responses/Unavailable'}
  /api/v1/context-collections/{id}/cancellation:
    post:
      summary: Persist cancellation intent for a collection requested by this principal
      description: >-
        Requires the original requester with current author permission. Accepts an
        empty JSON object and no query parameters; unknown or duplicate command fields are rejected. Repeated requests retain the
        first intent. Cancellation requested is distinct from observed Temporal
        cancellation; an unavailable workflow service cannot confirm that execution
        stopped. A receipt committed before cancellation remains historical fact.
      security: [{BearerAuth: []}, {BrowserCookie: []}]
      parameters:
        - $ref: '#/components/parameters/CollectionID'
        - $ref: '#/components/parameters/CollectionWorkspace'
        - $ref: '#/components/parameters/CollectionRepository'
        - $ref: '#/components/parameters/BrowserOrigin'
        - $ref: '#/components/parameters/CSRF'
      requestBody: {required: true, content: {application/json: {schema: {type: object, additionalProperties: false}}}}
      responses:
        '202': {description: Current collection with persisted cancellation intent or an earlier committed receipt, content: {application/json: {schema: {$ref: '#/components/schemas/Collection'}}}}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/CollectionStopped'}
        '500': {$ref: '#/components/responses/InternalError'}
        '503': {$ref: '#/components/responses/Unavailable'}
  /api/v1/changes/{id}/context-attachments:
    post:
      summary: Attach an inspected immutable collection receipt as a new package revision
      description: >-
        Requires current author permission and a receipt in the exact same
        workspace and canonical repository as the package. The command carries
        the inspected package revision, collection ID, and receipt digest.
        The service resolves trusted stored content and preserves other package
        fields transactionally. It does not collect newer source or carry approval
        onto the new revision. Unknown or duplicate command fields and query parameters are rejected; stale input requires
        explicit renewed inspection. A caller-written receipt ID is not trusted
        provenance by itself.
      security: [{BearerAuth: []}, {BrowserCookie: []}]
      parameters:
        - $ref: '#/components/parameters/ID'
        - $ref: '#/components/parameters/CollectionWorkspace'
        - $ref: '#/components/parameters/CollectionRepository'
        - $ref: '#/components/parameters/BrowserOrigin'
        - $ref: '#/components/parameters/CSRF'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [expectedRevision, collectionId, digest]
              properties:
                expectedRevision: {type: integer, format: int64, minimum: 1}
                collectionId: {type: string, pattern: '^[a-f0-9]{32}$'}
                digest: {type: string, pattern: '^[a-f0-9]{64}$', description: SHA-256 digest of the inspected stored receipt snapshot}
      responses:
        '201': {description: New unapproved package revision containing the inspected snapshot, content: {application/json: {schema: {$ref: '#/components/schemas/Package'}}}}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {description: Stale package revision or inspected receipt digest (revision_conflict), content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}
        '500': {$ref: '#/components/responses/InternalError'}
        '503': {$ref: '#/components/responses/Unavailable'}
...
  /api/v1/repository-graphs/{id}:
    parameters: [{$ref: '#/components/parameters/CollectionID'}, {$ref: '#/components/parameters/CollectionWorkspace'}, {$ref: '#/components/parameters/CollectionRepository'}, {$ref: '#/components/parameters/CSRF'}]
    get:
      summary: Inspect one immutable graph and explicit source gaps
      description: Requires current read permission on every source repository. Accepts no query parameters. Source text is retained in context receipts, not graph nodes. Freshness stays unknown.
      security: [{BearerAuth: []}, {BrowserCookie: []}]
      responses:
        '200': {description: Authorized graph snapshot, content: {application/json: {schema: {$ref: '#/components/schemas/RepositoryGraph'}}}}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '500': {$ref: '#/components/responses/InternalError'}
        '503': {$ref: '#/components/responses/Unavailable'}
...
  /api/v1/execution-profiles:
    get:
      summary: Discover enabled operator execution profiles for the selected workspace
      description: Requires explicit OIDC coordination and repository read access. Public profile/image
        digests are inspected before authorizing a plan. Secret paths and credentials are never returned.
        Discovery is capped at 100 entries with explicit truncation.
      security:
      - BearerAuth: []
      - BrowserCookie: []
      parameters:
      - $ref: '#/components/parameters/CollectionWorkspace'
      - $ref: '#/components/parameters/CollectionRepository'
      - $ref: '#/components/parameters/CSRF'
      responses:
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/Unavailable'
        '200':
          description: Bounded public execution profile catalog
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecutionProfilePage'
...
components:
  securitySchemes:
    BrowserCookie:
      type: apiKey
      in: cookie
      name: __Host-conductor-session
      description: Configured browser login only. Secure HttpOnly host-only session; cannot be combined with bearer or local credentials. Cookie mutations also require exact Origin and X-Conductor-CSRF. A supplied CSRF header is checked on reads, protecting stale tabs after account changes.
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: RFC 9068 RS256 JWT access token
      description: OIDC mode only; exact configured issuer and audience, at+jwt type, and signed required claims. ID tokens are not accepted.
    LocalActor:
      type: apiKey
      in: header
      name: X-Conductor-Actor
      description: Local development only. Not authentication and rejected in OIDC mode.
  responses:
...
    BadRequest: {description: Invalid command or query parameters, content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}
...
  parameters:
...
    ID: {name: id, in: path, required: true, schema: {type: string}}
    Actor: {name: X-Conductor-Actor, in: header, required: false, description: Required only in local mode; rejected in OIDC mode, schema: {type: string, maxLength: 128}}
...